refactor: stored redactions removed, works by mutating cold and hot storage now
