+++ lib/storage/middleware/save-storage-middleware.dart
@@ -11,9 +11,9 @@ import 'package:syphon/store/events/actions.dart';
 import 'package:syphon/store/events/messages/storage.dart';
 import 'package:syphon/store/events/reactions/actions.dart';
 import 'package:syphon/store/events/reactions/storage.dart';
+import 'package:syphon/store/events/receipts/actions.dart';
 import 'package:syphon/store/events/receipts/storage.dart';
 import 'package:syphon/store/events/redaction/actions.dart';
-import 'package:syphon/store/events/redaction/storage.dart';
 import 'package:syphon/store/index.dart';
 import 'package:syphon/store/media/actions.dart';
 import 'package:syphon/store/media/model.dart';
@@ -95,7 +95,6 @@ saveStorageMiddleware(Database? storageOld, StorageDatabase? storage) {
         break;
       case SaveRedactions:
         final _action = action as SaveRedactions;
-        saveRedactions(_action.redactions ?? [], storage: storageOld);
         saveMessagesRedacted(_action.redactions ?? [], storage: storage!);
         saveReactionsRedacted(_action.redactions ?? [], storage: storage);
         break;
