+++ lib/store/events/actions.dart
@@ -6,7 +6,6 @@ import 'package:syphon/global/print.dart';
 import 'package:syphon/storage/constants.dart';
 import 'package:syphon/storage/index.dart';
 import 'package:syphon/store/alerts/actions.dart';
-import 'package:syphon/store/events/ephemeral/m.read/model.dart';
 import 'package:syphon/store/events/messages/actions.dart';
 import 'package:syphon/store/events/messages/model.dart';
 import 'package:syphon/store/events/messages/storage.dart';
@@ -14,7 +13,6 @@ import 'package:syphon/store/events/model.dart';
 import 'package:syphon/store/index.dart';
 import 'package:syphon/store/rooms/actions.dart';
 import 'package:syphon/store/rooms/room/model.dart';
-import 'package:syphon/store/settings/models.dart';
 
 class ResetEvents {}
 
@@ -48,17 +46,6 @@ class AddMessages {
     this.outbox = const [],
     this.clear = false,
   });
-}
-
-class SetReceipts {
-  final String? roomId;
-  final Map<String, ReadReceipt>? receipts;
-  SetReceipts({this.roomId, this.receipts});
-}
-
-class LoadReceipts {
-  final Map<String, Map<String, ReadReceipt>> receiptsMap;
-  LoadReceipts({required this.receiptsMap});
 }
 
 ///
@@ -119,15 +106,6 @@ ThunkAction<AppState> addMessagesDecrypted({
       return store.dispatch(
         AddMessagesDecrypted(roomId: room.id, messages: messages, outbox: outbox),
       );
-    };
-
-ThunkAction<AppState> setReceipts({
-  Room? room,
-  Map<String, ReadReceipt>? receipts,
-}) =>
-    (Store<AppState> store) {
-      if (receipts!.isEmpty) return;
-      return store.dispatch(SetReceipts(roomId: room!.id, receipts: receipts));
     };
 
 ///
@@ -314,83 +292,6 @@ ThunkAction<AppState> selectReply({
   return (Store<AppState> store) async {
     final room = store.state.roomStore.rooms[roomId!]!;
     store.dispatch(SetRoom(room: room.copyWith(reply: message ?? Null)));
-  };
-}
-
-///
-/// Read Message Marker
-///
-/// Send Fully Read or just Read receipts bundled into
-/// one http call
-ThunkAction<AppState> sendReadReceipts({
-  Room? room,
-  Message? message,
-  bool readAll = true,
-}) {
-  return (Store<AppState> store) async {
-    try {
-      // Skip if typing indicators are disabled
-      if (store.state.settingsStore.readReceipts == ReadReceiptTypes.Off) {
-        return printInfo('[sendReadReceipts] read receipts disabled');
-      }
-
-      if (store.state.settingsStore.readReceipts == ReadReceiptTypes.Hidden) {
-        printInfo('[sendReadReceipts] read receipts hidden');
-      }
-
-      final data = await MatrixApi.sendReadReceipts(
-        protocol: store.state.authStore.protocol,
-        accessToken: store.state.authStore.user.accessToken,
-        homeserver: store.state.authStore.user.homeserver,
-        roomId: room!.id,
-        messageId: message!.id,
-        readAll: readAll,
-        hidden: store.state.settingsStore.readReceipts == ReadReceiptTypes.Hidden,
-      );
-
-      if (data['errcode'] != null) {
-        throw data['error'];
-      }
-
-      printInfo('[sendReadReceipts] sent ${message.id} $data');
-    } catch (error) {
-      printInfo('[sendReadReceipts] failed $error');
-    }
-  };
-}
-
-///
-/// Read Message Marker
-///
-/// Send Fully Read or just Read receipts bundled into
-/// one http call
-ThunkAction<AppState> sendTyping({
-  String? roomId,
-  bool? typing = false,
-}) {
-  return (Store<AppState> store) async {
-    try {
-      // Skip if typing indicators are disabled
-      if (!store.state.settingsStore.typingIndicatorsEnabled) {
-        printInfo('[sendTyping] typing indicators disabled');
-        return;
-      }
-
-      final data = await MatrixApi.sendTyping(
-        protocol: store.state.authStore.protocol,
-        accessToken: store.state.authStore.user.accessToken,
-        homeserver: store.state.authStore.user.homeserver,
-        roomId: roomId,
-        userId: store.state.authStore.user.userId,
-        typing: typing,
-      );
-
-      if (data['errcode'] != null) {
-        throw data['error'];
-      }
-    } catch (error) {
-      printError('[sendTyping] $error');
-    }
   };
 }
