+++ lib/views/home/home-screen.dart
@@ -759,7 +759,6 @@ class _Props extends Equatable {
   final Map<String, List<Message>> messages;
   final Map<String, List<Message>> decrypted;
 
-  final Function onDebug;
   final Function onLeaveChat;
   final Function onDeleteChat;
   final Function onArchiveChat;
@@ -782,7 +781,6 @@ class _Props extends Equatable {
     required this.fabType,
     required this.fabLocation,
     required this.roomTypeBadgesEnabled,
-    required this.onDebug,
     required this.onLeaveChat,
     required this.onDeleteChat,
     required this.onSelectHelp,
@@ -854,9 +852,6 @@ class _Props extends Equatable {
               throw 'Could not launch ${Values.openHelpUrl}';
             }
           } catch (error) {}
-        },
-        onDebug: () async {
-          printDebug('[onDebug] trigged debug function @ home');
         },
       );
 }
